+++ packages/core/src/di/core-tokens.ts
@@ -25,6 +25,7 @@ export const pluginsByKind = 'pluginsByKind';
 export const validationSchema = 'validationSchema';
 export const optionsValidator = 'optionsValidator';
 export const requireFromCwd = 'requireFromCwd';
+export const resolveFromCwd = 'resolveFromCwd';
 export const fs = 'fs';
 export const testCoverage = 'testCoverage';
 export const incrementalDiffer = 'incrementalDiffer';