+++ packages/core/src/initializer/initializer-tokens.ts
@@ -1,7 +1,7 @@
 export const restClientNpmSearch = 'restClientNpmSearch';
 export const restClientNpm = 'restClientNpm';
 export const npmClient = 'npmClient';
-export const strykerPresets = 'strykerPresets';
+export const customInitializers = 'strykerPresets';
 export const configWriter = 'configWriter';
 export const gitignoreWriter = 'gitignoreWriter';
 export const inquirer = 'inquirer';